# swagger/admin.yaml
swagger: "2.0"
info:
  description: "机器人管理后台"
  version: "1.0.0"
  title: "Robot admin Api"
  contact:
    email: "372290258@qq.com"
host: “robot.weima123.cn”
basePath: "/admin"
...
schemes:
- "http"
paths:
  /auth/login:
    post:
      x-swagger-router-controller: "auth"
      tags:
      - "auth"
      summary: "管理员登录"
      description: "管理员登录管理系统"
      # operationId: "login"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "body"
        in: "body"
        description: "登录信息"
        required: true
        schema:
          $ref: "#/definitions/Login"
      responses:
        405:
          description: "Invalid operation"
          schema:
            $ref: "#/definitions/ApiResponse"
        200:
          description: "successful operation"
          schema:
            type: "object"
            properties:
              status:
                type: "integer"
              result:
                $ref: "#/definitions/Manager"
  /auth/logout:
    get:
      x-swagger-router-controller: "auth"
      tags:
      - "auth"
      summary: "管理员退出"
      description: "管理员退出管理系统"
      # operationId: "logout"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      security:
      - cookie: []
  /managers:
    get:
      x-swagger-router-controller: "manager"
      tags:
      - "manager"
      summary: "管理员管理"
      description: "获取管理员列表"
      # operationId: "search"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "query"
        required: true
        type: "integer"
        description: "查询页号"
      - name: "number"
        in: "query"
        required: true
        type: "integer"
        description: "每页数量"
      - name: "keyword"
        in: "query"
        required: true
        type: "string"
        description: "关键词"
      - name: "order"
        in: "query"
        required: true
        type: "string"
        description: "排序字段"
      - name: "direction"
        in: "query"
        required: true
        type: "string"
        description: "排序方向：asc，decs"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "object"
            properties:
              status:
                type: "integer"
              result:
                type: "object"
                properties:
                  count:
                    type: "integer"
                  list:
                    type: "array"
                    items:
                      $ref: "#/definitions/Manager"
      security:
      - cookie: []
  /manager:
    post:
      x-swagger-router-controller: "manager"
      tags:
      - "manager"
      summary: "添加管理员"
      description: "添加系统管理员"
      # # operationId: "add"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "body"
        in: "body"
        required: true
        schema:
          $ref: "#/definitions/ManagerInfo"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/IdResponse"
      security:
      - cookie: []
  /manager/{manager_id}:
    get:
      x-swagger-router-controller: "manager"
      tags:
      - "manager"
      summary: "查看管理员"
      description: "查看单个管理员"
      # operationId: "get"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "manager_id"
        in: "path"
        description: "管理员id"
        required: true
        type: "integer"
      responses:
        405:
          description: "Invalid operation"
          schema:
            $ref: "#/definitions/ApiResponse"
        200:
          description: "successful operation"
          schema:
            type: "object"
            properties:
              status:
                type: "integer"
              result:
                $ref: "#/definitions/Manager"
      security:
      - cookie: []
    put:
      x-swagger-router-controller: "manager"
      tags:
      - "manager"
      summary: "修改管理员"
      description: "修改系统管理员"
      # operationId: "edit"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "manager_id"
        in: "path"
        description: "管理员id"
        required: true
        type: "integer"
      - name: "body"
        in: "body"
        required: true
        schema:
          $ref: "#/definitions/ManagerInfo"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      security:
      - cookie: []
    delete:
      x-swagger-router-controller: "manager"
      tags:
      - "manager"
      summary: "删除管理员"
      description: "删除系统管理员"
      # operationId: "remove"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "manager_id"
        in: "path"
        description: "管理员id"
        required: true
        type: "integer"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      security:
      - cookie: []
  /robots:
    get:
      x-swagger-router-controller: "robot"
      tags:
      - "robot"
      summary: "机器人管理"
      description: "获取机器人列表"
      # operationId: "search"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "query"
        required: true
        type: "integer"
        description: "查询页号"
      - name: "number"
        in: "query"
        required: true
        type: "integer"
        description: "每页数量"
      - name: "keyword"
        in: "query"
        required: true
        type: "string"
        description: "关键词"
      - name: "order"
        in: "query"
        required: true
        type: "string"
        description: "排序字段"
      - name: "direction"
        in: "query"
        required: true
        type: "string"
        description: "排序方向：asc，decs"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "object"
            properties:
              status:
                type: "integer"
              result:
                type: "object"
                properties:
                  count:
                    type: "integer"
                  list:
                    type: "array"
                    items:
                      $ref: "#/definitions/Robot"
      security:
      - cookie: []
  /robot:
    post:
      x-swagger-router-controller: "robot"
      tags:
      - "robot"
      summary: "添加机器人"
      description: "添加机器人"
      # operationId: "add"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "content"
        in: "formData"
        required: true
        description: "机器人标识码文本，一行一个"
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      security:
      - cookie: []
  /robot/{robot_id}:
    get:
      x-swagger-router-controller: "robot"
      tags:
      - "robot"
      summary: "查看机器人"
      description: "查看单个机器人"
      # operationId: "get"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "robot_id"
        in: "path"
        description: "机器人id"
        required: true
        type: "integer"
      responses:
        405:
          description: "Invalid operation"
          schema:
            $ref: "#/definitions/ApiResponse"
        200:
          description: "successful operation"
          schema:
            type: "object"
            properties:
              status:
                type: "integer"
              result:
                $ref: "#/definitions/Robot"
      security:
      - cookie: []
    put:
      x-swagger-router-controller: "robot"
      tags:
      - "robot"
      summary: "修改机器人"
      description: "修改机器人信息"
      # operationId: "edit"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "robot_id"
        in: "path"
        description: "机器人id"
        required: true
        type: "integer"
      - name: "body"
        in: "body"
        required: true
        schema:
          $ref: "#/definitions/RobotInfo"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      security:
      - cookie: []
  /users:
    get:
      x-swagger-router-controller: "user"
      tags:
      - "user"
      summary: "用户管理"
      description: "获取用户列表"
      # operationId: "search"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "query"
        required: true
        type: "integer"
        description: "查询页号"
      - name: "number"
        in: "query"
        required: true
        type: "integer"
        description: "每页数量"
      - name: "keyword"
        in: "query"
        required: true
        type: "string"
        description: "关键词"
      - name: "order"
        in: "query"
        required: true
        type: "string"
        description: "排序字段"
      - name: "direction"
        in: "query"
        required: true
        type: "string"
        description: "排序方向：asc，decs"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "object"
            properties:
              status:
                type: "integer"
              result:
                type: "object"
                properties:
                  count:
                    type: "integer"
                  list:
                    type: "array"
                    items:
                      $ref: "#/definitions/User"
      security:
      - cookie: []
  /user/{user_id}:
    get:
      x-swagger-router-controller: "user"
      tags:
      - "user"
      summary: "查看用户"
      description: "查看单个用户"
      # operationId: "get"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "user_id"
        in: "path"
        description: "用户id"
        required: true
        type: "integer"
      responses:
        405:
          description: "Invalid operation"
          schema:
            $ref: "#/definitions/ApiResponse"
        200:
          description: "successful operation"
          schema:
            type: "object"
            properties:
              status:
                type: "integer"
              result:
                $ref: "#/definitions/User"
      security:
      - cookie: []
    put:
      x-swagger-router-controller: "user"
      tags:
      - "user"
      summary: "修改用户"
      description: "修改用户信息"
      # operationId: "edit"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "user_id"
        in: "path"
        description: "用户id"
        required: true
        type: "integer"
      - name: "body"
        in: "body"
        required: true
        schema:
          $ref: "#/definitions/UserInfo"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      security:
      - cookie: []
  /banners:
    get:
      x-swagger-router-controller: "banner"
      tags:
      - "base"
      summary: "Banner管理"
      description: "获取Banner列表"
      # operationId: "search"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "query"
        required: true
        type: "integer"
        description: "查询页号"
      - name: "number"
        in: "query"
        required: true
        type: "integer"
        description: "每页数量"
      - name: "keyword"
        in: "query"
        required: true
        type: "string"
        description: "关键词"
      - name: "order"
        in: "query"
        required: true
        type: "string"
        description: "排序字段"
      - name: "direction"
        in: "query"
        required: true
        type: "string"
        description: "排序方向：asc，decs"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "object"
            properties:
              status:
                type: "integer"
              result:
                type: "object"
                properties:
                  count:
                    type: "integer"
                  list:
                    type: "array"
                    items:
                      $ref: "#/definitions/Banner"
      security:
      - cookie: []
  /banner:
    post:
      x-swagger-router-controller: "base"
      tags:
      - "base"
      summary: "添加Banner"
      description: "添加Banner"
      # operationId: "add"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "body"
        in: "body"
        required: true
        schema:
          $ref: "#/definitions/BannerInfo"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/IdResponse"
      security:
      - cookie: []
  /banner/{banner_id}:
    get:
      x-swagger-router-controller: "banner"
      tags:
      - "base"
      summary: "查看Banner"
      description: "查看单个Banner信息"
      # operationId: "get"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "banner_id"
        in: "path"
        description: "Banner：id"
        required: true
        type: "integer"
      responses:
        405:
          description: "Invalid operation"
          schema:
            $ref: "#/definitions/ApiResponse"
        200:
          description: "successful operation"
          schema:
            type: "object"
            properties:
              status:
                type: "integer"
              result:
                $ref: "#/definitions/Banner"
      security:
      - cookie: []
    put:
      x-swagger-router-controller: "banner"
      tags:
      - "base"
      summary: "修改Banner"
      description: "修改Banner信息"
      # operationId: "edit"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "banner_id"
        in: "path"
        description: "Banner：id"
        required: true
        type: "integer"
      - name: "body"
        in: "body"
        required: true
        schema:
          $ref: "#/definitions/BannerInfo"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      security:
      - cookie: []
    delete:
      x-swagger-router-controller: "banner"
      tags:
      - "base"
      summary: "删除Banner"
      description: "删除Banner信息"
      # operationId: "remove"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "banner_id"
        in: "path"
        description: "Banner：id"
        required: true
        type: "integer"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      security:
      - cookie: []
  /banner/{banner_id}/list:
    get:
      x-swagger-router-controller: "banner"
      tags:
      - "base"
      summary: "查看Banner下关联信息"
      description: "查看单个Banner信息"
      # operationId: "getRelation"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "banner_id"
        in: "path"
        description: "Banner：id"
        required: true
        type: "integer"
      - name: "page"
        in: "query"
        required: true
        type: "integer"
        description: "查询页号"
      - name: "number"
        in: "query"
        required: true
        type: "integer"
        description: "每页数量"
      - name: "order"
        in: "query"
        required: true
        type: "string"
        description: "排序字段"
      - name: "direction"
        in: "query"
        required: true
        type: "string"
        description: "排序方向：asc，decs"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "object"
            properties:
              status:
                type: "integer"
              result:
                type: "object"
                properties:
                  count:
                    type: "integer"
                  list:
                    type: "array"
                    items:
                      $ref: "#/definitions/BannerRelation"
      security:
      - cookie: []
    post:
      x-swagger-router-controller: "banner"
      tags:
      - "base"
      summary: "添加Banner关系"
      description: "添加Banner的关联关系"
      # operationId: "addRelation"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "banner_id"
        in: "path"
        description: "Banner：id"
        required: true
        type: "integer"
      - name: "module"
        in: "formData"
        required: true
        description: "对应模块"
        type: "string"
      - name: "ids"
        in: "formData"
        required: true
        description: "对应模块id"
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      security:
      - cookie: []
  /sound/albums:
    get:
      x-swagger-router-controller: "sound"
      tags:
      - "sound"
      summary: "磨耳朵专辑管理"
      description: "获取磨耳朵专辑列表"
      # operationId: "searchAlbum"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "query"
        required: true
        type: "integer"
        description: "查询页号"
      - name: "number"
        in: "query"
        required: true
        type: "integer"
        description: "每页数量"
      - name: "keyword"
        in: "query"
        required: true
        type: "string"
        description: "关键词"
      - name: "order"
        in: "query"
        required: true
        type: "string"
        description: "排序字段"
      - name: "direction"
        in: "query"
        required: true
        type: "string"
        description: "排序方向：asc，decs"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "object"
            properties:
              status:
                type: "integer"
              result:
                type: "object"
                properties:
                  count:
                    type: "integer"
                  list:
                    type: "array"
                    items:
                      $ref: "#/definitions/SoundAlbum"
      security:
      - cookie: []
  /sound/albums_ids:
    get:
      x-swagger-router-controller: "sound"
      tags:
      - "sound"
      summary: "磨耳朵专辑管理"
      description: "获取磨耳朵专辑列表"
      # operationId: "searchAlbum"
      produces:
      - "application/json"
      parameters:
      - name: "ids"
        in: "query"
        required: true
        type: "string"
        description: "获取具体id"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "object"
            properties:
              status:
                type: "integer"
              result:
                type: "object"
                properties:
                  count:
                    type: "integer"
                  list:
                    type: "array"
                    items:
                      $ref: "#/definitions/SoundAlbum"
      security:
      - cookie: []
  /sound/album:
    post:
      x-swagger-router-controller: "sound"
      tags:
      - "sound"
      summary: "添加磨耳朵专辑"
      description: "添加磨耳朵专辑"
      # operationId: "addAlbum"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "body"
        in: "body"
        required: true
        schema:
          $ref: "#/definitions/SoundAlbumInfo"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/IdResponse"
      security:
      - cookie: []
...
  /sound/album/{album_id}/categorys:
    get:
      x-swagger-router-controller: "sound"
      tags:
      - "sound"
      summary: "磨耳朵专辑分类"
      description: "获取磨耳朵专辑分类"
      # operationId: "getAlbumCategory"
      produces:
      - "application/json"
      parameters:
      - name: "album_id"
        in: "path"
        description: "专辑id"
        required: true
        type: "integer"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "object"
            properties:
              status:
                type: "integer"
              result:
                type: "object"
                properties:
                  count:
                    type: "integer"
                  list:
                    type: "array"
                    items:
                      $ref: "#/definitions/SoundCategory"
      security:
      - cookie: []
...
  /sound/voice:
    post:
      x-swagger-router-controller: "sound"
      tags:
      - "sound"
      summary: "添加磨耳朵音频"
      description: "添加磨耳朵音频"
      # operationId: "addVoice"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "body"
        in: "body"
        required: true
        schema:
          $ref: "#/definitions/SoundVoiceInfo"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/IdResponse"
      security:
      - cookie: []
...
securityDefinitions:
  cookie:
    type: apiKey
    in: header
    name: Cookie
definitions:
  Login:
    type: "object"
    properties:
      username:
        type: "string"
      password:
        type: "string"
  Manager:
    type: "object"
    properties:
      id:
        type: "integer"
      username:
        type: "string"
      status:
        type: "integer"
      last_time:
        type: "string"
  ManagerInfo:
    type: "object"
    properties:
      username:
        type: "string"
      password:
        type: "string"
  Robot:
    type: "object"
    properties:
      id:
        type: "integer"
      code:
        type: "string"
        description: "机器人识别码"
      status:
        type: "integer"
      alive:
        type: "boolean"
      version:
        type: "string"
      avatar:
        type: "string"
      nickname:
        type: "string"
      birthday:
        type: "string"
      gender:
        type: "integer"
  RobotInfo:
    type: "object"
    properties:
      status:
        type: "integer"
      avatar:
        type: "string"
      nickname:
        type: "string"
      birthday:
        type: "string"
      gender:
        type: "integer"
  User:
    type: "object"
    properties:
      id:
        type: "integer"
      username:
        type: "string"
      mobile:
        type: "string"
      area:
        type: "string"
      avatar:
        type: "string"
      nickname:
        type: "string"
      wechat:
        type: "string"
      role:
        type: "integer"
...
  IdResponse:
    type: "object"
    properties:
      status:
        type: "integer"
      result:
        type: "object"
        properties:
          id:
            type: "integer"
            description: "id"
  ApiResponse:
    type: "object"
    properties:
      status:
        type: "integer"
      result:
        type: "object"
      message:
        type: "string"